code/requirements.txt
# Python dependencies for Enoch Calendar Analysis
# Supplementary code for manuscript: "Phase-Locking in the 364-Day Calendar"
#
# Install with: pip install -r requirements.txt
# Tested with Python 3.10+

# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Ephemeris calculations (Swiss Ephemeris bindings)
pyswisseph>=2.10.3.2

# Environment variables
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
nbconvert>=7.0.0  # For programmatic notebook execution

# Optional: for better performance
# numba>=0.57.0
